tailorcv-backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
supabase==2.0.0
python-multipart==0.0.6
werkzeug==2.3.7
pdflatex==0.1.3
jinja2==3.1.2
linkedin-api==2.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
pillow==10.0.1
gunicorn==21.2.0